Move file_browser components into a subdirectory.
