+++ app/javascript/components/file_browser/directory_contents.vue
