+++ app/javascript/components/file_browser/directory_picker.vue
