+++ app/javascript/components/file_browser/embedded_file_browser.vue
