+++ app/javascript/components/file_browser/file_browser.vue
