+++ app/javascript/components/file_browser/input_path_selector.vue
